Refactor examples app dir structure by separates the main app (router) to it is own dir
