+++ examples-preview/public/index.html
@@ -9,6 +9,9 @@
     <div id="app-root">
     </div>
   </body>
-  <script src="/bundle.js"></script>
+
+  <!-- examples-preview-app.js is generated by webpack -->
+  <!-- see webpack.config.js file -->
+  <script src="/examples-preview-app.js"></script>
 </html>
